crypto-bot-mvp/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client for API requests
aiohttp==3.9.1
httpx==0.25.2

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
ta==0.10.2

# Machine learning and sentiment analysis
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Redis for caching and data persistence
redis==5.0.1

# Database and data storage
sqlalchemy==2.0.23
alembic==1.13.1

# WebSocket support
websockets==12.0

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3